refactor(components): remove ce extension
for style and js file
