+++ webcomponents/src/components/PixLabel.scss
