+++ webcomponents/src/components/PixTag.ce.vue
@@ -1,5 +1,6 @@
 <script setup>
 import './PixTag.scss';
+
 defineProps({
   color: String,
 });
