+++ webcomponents/src/components/PixToggleButton.scss
@@ -1,0 +1,62 @@
+@use '@assets/pix-design-tokens/index.scss';
+
+.pix-toggle-button {
+  display: inline-flex;
+  flex-direction: column;
+  gap: var(--pix-spacing-1x);
+
+  &--inline {
+    flex-direction: row;
+    gap: var(--pix-spacing-2x);
+    align-items: center;
+  }
+
+  &__button {
+    @extend %pix-body-s;
+
+    width: fit-content;
+    padding: var(--pix-spacing-1x);
+    font-weight: var(--pix-font-bold);
+    line-height: 1.572;
+    background: var(--pix-neutral-20);
+    border: 1px solid var(--pix-neutral-500);
+    border-radius: 4px;
+
+    &--icon {
+      font-size: 0;
+    }
+  }
+
+  &__view-a,
+  &__view-b {
+    display: inline-block;
+    padding: var(--pix-spacing-1x) var(--pix-spacing-2x);
+    border-radius: 4px;
+  }
+
+  &__view-a {
+    color: var(--pix-neutral-800);
+    font-weight: var(--pix-font-normal);
+  }
+
+  &__view-b {
+    color: var(--pix-neutral-20);
+    background-color: var(--pix-neutral-800);
+  }
+
+  &--pressed {
+    .pix-toggle-button {
+      &__view-a {
+        color: var(--pix-neutral-20);
+        font-weight: inherit;
+        background-color: var(--pix-neutral-800);
+      }
+
+      &__view-b {
+        color: var(--pix-neutral-800);
+        font-weight: var(--pix-font-normal);
+        background-color: transparent;
+      }
+    }
+  }
+}
