+++ webcomponents/src/main.js
@@ -1,3 +1,3 @@
 import './components/PixTag.ce.js';
 import './components/PixLabel.js';
-import './components/PixToggleButton.ce.js';
+import './components/PixToggleButton.js';
